+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * Renamed "Solar Panel" to "Photovoltaic Cell" to avoid confusions with solar generators
 * Photovoltaic Cells can no longer be placed
 * (API) Removed deprecated "SlimefunBlockHandler"
+* Removed Automated Crafting Chamber
 
 #### Fixes
 * Fixed #2987